EEA wrapping error fixed

+++ labrador/src/zq.rs
@@ -55,10 +55,10 @@ impl Zq {
     /// Calculates the modular inverse of an element using the Extended Euclidean Algorithm.
     /// Returns `None` if no inverse exists.
     pub fn inv(self) -> Option<Self> {
-        let mut a = self.value;
-        let mut x0: u32 = 0;
-        let mut x1: u32 = 1;
-        let mut q: u32 = 4294967291; // modulus 2^32 - 5
+        let mut a = self.value as u128;
+        let mut x0: u128 = 0;
+        let mut x1: u128 = 1;
+        let mut q: u128 = 4294967291; // modulus 2^32 - 5
         let original_q = q;
 
         while a != 0 {
@@ -82,7 +82,7 @@ impl Zq {
         let result =
             (x1.wrapping_rem(original_q).wrapping_add(original_q)).wrapping_rem(original_q);
 
-        Some(Zq::new(result))
+        Some(Zq::new((result) as u32))
     }
 }
